Classes moved to the newly created Computation and Computation.Constraint sub namespaces of Machine namespace.
