+++ src/TuringMachine/TuringMachine/Extensions/ExceptionCustomizer/ExceptionToStringCustomizer.cs
@@ -8,6 +8,13 @@ namespace TuringMachine.Extensions.ExceptionCustomizer
     /// </summary>
     internal static class ExceptionToStringCustomizer
     {
+        /// <summary>
+        /// Adds custom text to the specified exception's <see cref="Exception.ToString"/> result.
+        /// </summary>
+        /// <typeparam name="TException">Type of the exception.</typeparam>
+        /// <param name="exception">Exception.</param>
+        /// <param name="customTexts">Texts that are inserted between the message and stacktrace parts.</param>
+        /// <returns><see cref="string"/> that contains the typename, message, custom texts and the stacktrace in this order.</returns>
         public static string CustomizeToString<TException>(this TException exception, params string[] customTexts) 
             where TException : Exception
         {
