+++ src/TuringMachine/TuringMachine/Machine/Computation/ComputationAbortedEventArgs.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace TuringMachine.Machine
+namespace TuringMachine.Machine.Computation
 {
     /// <summary>
     /// Provides data for the event that is raised when a computation has been aborted.
