+++ src/TuringMachine/TuringMachine/Machine/Computation/IManualComputation.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
-using TuringMachine.Machine.ComputationConstraint;
+using TuringMachine.Machine.Computation.Constraint;
 
-namespace TuringMachine.Machine
+namespace TuringMachine.Machine.Computation
 {
     /// <summary>
     /// Defines methods for controlling manual computations.
