+++ src/TuringMachine/UnitTest/UnitTests/ComputationConstraint/TimeLimitConstraintTests.cs
@@ -1,7 +1,7 @@
 ﻿using Moq;
 using System;
-using TuringMachine.Machine;
-using TuringMachine.Machine.ComputationConstraint;
+using TuringMachine.Machine.Computation;
+using TuringMachine.Machine.Computation.Constraint;
 using Xunit;
 
 namespace TuringMachine.Tests.UnitTests.ComputationConstraint
